Brought everyone upto date in app.java and Sql.java
