+++ src/main/java/com/napier/sem/App.java
@@ -2,7 +2,6 @@ package com.napier.sem;
 
 import com.napier.sem.db.DbConnection;
 import com.napier.sem.report.Reports;
-import com.napier.sem.sql.AdditionalReports;
 import com.napier.sem.sql.Sql;
 import com.napier.sem.world.City;
 import com.napier.sem.world.Country;
@@ -15,7 +14,7 @@ import java.util.ArrayList;
  * Project Authors: Colin, Cameron, Luke, Del
  * Date last modified: 22/02/2021
  * Purpose of program: To run SQL queries to satisfy the given requirements
- * Last modified by: Cameron
+ * Last modified by: Colin
  */
 
 public class App
@@ -23,124 +22,103 @@ public class App
     public static void main(String[] args)
     {
         // Status update
-        System.out.println("Added Con & tidied up tested full - working");
+        System.out.println("Brought everyone upto date in app");
 
         // Database connection setup
         DbConnection dbc = new DbConnection();
         Connection con = dbc.getConnection();
-        Sql query = new Sql();
-        Reports report = new Reports();
 
-        // Return all countries in the world to countries1 object
-        ArrayList<Country> countries1 = query.getAllCountriesInWorld(con);
-        // Display report from countries1
-        report.printCountriesInWorld(countries1);
+        /************** COUNTRIES FUNCTIONS ***************** Author Luke */
+/*
+        // Report 1 - Return all countries in the world to allCountriesInWorld object
+        ArrayList<Country> allCountriesInWorld = Sql.getAllCountriesInWorld(con);
+        // Display report from allCountriesInWorld
+        Reports.printCountriesInWorld(allCountriesInWorld);
 
-        // Return all countries in a continent to countries2 object
-        ArrayList<Country> countries2 = query.getAllCountriesInContinent(con, "Europe");
-        // Display report from countries2
-        report.printCountriesInContinent(countries2);
+        // Report 2 - Return all countries in a continent to allCountriesInContinent object
+        ArrayList<Country> allCountriesInContinent = Sql.getAllCountriesInContinent(con, "Europe");
+        // Display report from allCountriesInContinent
+        Reports.printCountriesInContinent(allCountriesInContinent);
 
-        // Return all countries in a region to countries3 object
-        ArrayList<Country> countries3 = query.getAllCountriesByRegion(con, "Caribbean");
-        // Display report from countries3
-        report.printCountriesInRegion(countries3);
+        // Report 3 - Return all countries in a region to allCountriesByRegion object
+        ArrayList<Country> allCountriesByRegion = Sql.getAllCountriesByRegion(con, "Caribbean");
+        // Display report from allCountriesByRegion
+        Reports.printCountriesInRegion(allCountriesByRegion);
 
-        // Return the top N countries in the world to countries4 object
-        ArrayList<Country> countries4 = query.getTopPopulatedCountriesInWorld(con, 10);
-        // Display report from countries4
-        report.printCountriesInWorld(countries4);
+        // Report 4 - Return the top N countries in the world to topPopulatedCountriesInWorld object
+        ArrayList<Country> topPopulatedCountriesInWorld = Sql.getTopPopulatedCountriesInWorld(con, 10);
+        // Display report from topPopulatedCountriesInWorld
+        Reports.printCountriesInWorld(topPopulatedCountriesInWorld);
 
-        // Return the top N countries in a continent to countries5 object
-        ArrayList<Country> countries5 = query.getTopPopulatedCountriesInContinent(con, "Europe", 10);
-        // Display report from countries5
-        report.printCountriesInContinent(countries5);
+        // Report 5 - Return the top N countries in a continent to topPopulatedCountriesInContinent object
+        ArrayList<Country> topPopulatedCountriesInContinent = Sql.getTopPopulatedCountriesInContinent(con, "Europe", 10);
+        // Display report from topPopulatedCountriesInContinent
+        Reports.printCountriesInContinent(topPopulatedCountriesInContinent);
 
-        // Return the top N countries in a region to countries6 object
-        ArrayList<Country> countries6 = query.getTopPopulatedCountriesInRegion(con, "Caribbean", 10);
-        // Display report from countries6
-        report.printCountriesInRegion(countries6);
+        // Report 6 - Return the top N countries in a region to topPopulatedCountriesInRegion object
+        ArrayList<Country> topPopulatedCountriesInRegion = Sql.getTopPopulatedCountriesInRegion(con, "Caribbean", 10);
+        // Display report from topPopulatedCountriesInRegion
+        Reports.printCountriesInRegion(topPopulatedCountriesInRegion);
+*/
+        /************** CITIES FUNCTIONS ***************** Author Colin */
 
-        // Return all cities in world to cities object
+        // Report 7 - Return all cities in world to allCitiesInWorld object
         ArrayList<City> allCitiesInWorld = Sql.getAllCitiesInWorld(con);
-        // Display report from cities
+        // Display report from allCitiesInWorld
         assert allCitiesInWorld != null;
         Reports.printCities(allCitiesInWorld);
+/*
+        // Report 8 - Return all cities in a continent to allCitiesInContinent object
+        ArrayList<City> allCitiesInContinent = Sql.getAllCitiesInContinent(con, "Asia");
+        // Display report from allCitiesInContinent
+        Reports.printCities(allCitiesInContinent);
 
-        // Return all cities in a continent to cities8 object
-        Sql query8 = new Sql();
-        ArrayList<City> cities8 = query8.getAllCitiesInContinent(con, "Asia");
+        // Report 9 - Return all cities in a region to allCitiesInRegion object
+        ArrayList<City> allCitiesInRegion = Sql.getAllCitiesInRegion(con, "Caribbean");
+        // Display report from allCitiesInRegion
+        Reports.printCities(allCitiesInRegion);
 
-        // Display report from cities8
-        Reports report8 = new Reports();
-        //report8.printCities(cities8);
+        // Report 10 - Return all cities in a region to allCitiesInCountry object
+        ArrayList<City> allCitiesInCountry = Sql.getAllCitiesInCountry(con, "Brazil");
+        // Display report from allCitiesInCountry
+        Reports.printCities(allCitiesInCountry);
 
-        // Return all cities in a region to cities9 object
-        Sql query9 = new Sql();
-        ArrayList<City> cities9 = query9.getAllCitiesInRegion(con, "Caribbean");
+        // Report 11 - Return all cities in a district to allCitiesInDistrict object
+        ArrayList<City> allCitiesInDistrict = Sql.getAllCitiesInDistrict(con, "Scotland");
+        // Display report from allCitiesInDistrict
+        Reports.printCities(allCitiesInDistrict);
 
-        // Display report from cities9
-        Reports report9 = new Reports();
-        //report9.printCities(cities9);
+        // Report 12 - Return top N populated cities in the world to topPopulatedCitiesInWorld object
+        ArrayList<City> topPopulatedCitiesInWorld = Sql.getTopPopulatedCitiesInWorld(con, 10);
+        // Display report from topPopulatedCitiesInWorld
+        Reports.printCities(topPopulatedCitiesInWorld);
 
-        // Return all cities in a region to cities10 object
-        Sql query10 = new Sql();
-        ArrayList<City> cities10 = query10.getAllCitiesInCountry(con, "Brazil");
+        // Report 13 - Return top N populated cities in a continent to topPopulatedCitiesInContinent object
+        ArrayList<City> topPopulatedCitiesInContinent = Sql.getTopPopulatedCitiesInContinent(con, 10, "Asia");
+        // Display report from topPopulatedCitiesInContinent
+        Reports.printCities(topPopulatedCitiesInContinent);
 
-        // Display report from cities10
-        Reports report10 = new Reports();
-        //report10.printCities(cities10);
+        // Report 14 - Return top N populated cities in a region to topPopulatedCitiesInRegion object
+        ArrayList<City> topPopulatedCitiesInRegion = Sql.getTopPopulatedCitiesInRegion(con, 10, "Caribbean");
+        // Display report from topPopulatedCitiesInRegion
+        Reports.printCities(topPopulatedCitiesInRegion);
 
-        // Return all cities in a district to cities11 object
-        Sql query11 = new Sql();
-        ArrayList<City> cities11 = query11.getAllCitiesInDistrict(con, "Scotland");
+        // Report 15 - Return top N populated cities in the world to topPopulatedCitiesInCountry object
+        ArrayList<City> topPopulatedCitiesInCountry = Sql.getTopPopulatedCitiesInCountry(con, 10, "Brazil");
+        // Display report from topPopulatedCitiesInCountry
+        Reports.printCities(topPopulatedCitiesInCountry);
 
-        // Display report from cities11
-        Reports report11 = new Reports();
-        //report11.printCities(cities11);
-
-        // Return top N populated cities in the world to cities12 object
-        Sql query12 = new Sql();
-        ArrayList<City> cities12 = query12.getTopPopulatedCitiesInWorld(con, 10);
-
-        // Display report from cities12
-        Reports report12 = new Reports();
-        //report12.printCities(cities12);
-
-        // Return top N populated cities in a continent to cities13 object
-        Sql query13 = new Sql();
-        ArrayList<City> cities13 = query13.getTopPopulatedCitiesInContinent(con, 10, "Asia");
-
-        // Display report from cities13
-        Reports report13 = new Reports();
-        //report13.printCities(cities13);
-
-        // Return top N populated cities in a region to cities14 object
-        Sql query14 = new Sql();
-        ArrayList<City> cities14 = query14.getTopPopulatedCitiesInRegion(con, 10, "Caribbean");
-
-        // Display report from cities14
-        Reports report14 = new Reports();
-        //report14.printCities(cities14);
-
-        // Return top N populated cities in the world to cities15 object
-        Sql query15 = new Sql();
-        ArrayList<City> cities15 = query15.getTopPopulatedCitiesInCountry(con, 10, "Brazil");
-
-        // Display report from cities15
-        Reports report15 = new Reports();
-        //report15.printCities(cities15);
-
-        // Return top N populated cities in the world to cities16 object
-        Sql query16 = new Sql();
-        ArrayList<City> cities16 = query16.getTopPopulatedCitiesInDistrict(con, 10, "Scotland");
-
-        // Display report from cities16
-        Reports report16 = new Reports();
-        //report16.printCities(cities16);
+        // Report 16 - Return top N populated cities in the world to topPopulatedCitiesInDistrict object
+        ArrayList<City> topPopulatedCitiesInDistrict = Sql.getTopPopulatedCitiesInDistrict(con, 10, "Scotland");
+        // Display report from topPopulatedCitiesInDistrict
+        Reports.printCities(topPopulatedCitiesInDistrict);
+*/
+        /************** CAPITAL CITIES FUNCTIONS ***************** Author Del */
 
 
-        // Additional Information Reports //
+
+
+        /************** ADDITIONAL FUNCTIONS ***************** Author Cameron */
         // Report 1:
         // Get population of the world
         // Long totalPopulation = AdditionalReports.getWorldPopulation(con);
@@ -151,52 +129,52 @@ public class App
         // Set continent to get population of
         // 'Asia','Europe','North America','Africa','Oceania','Antarctica','South America'
         // String continent = "Europe";
-        // Long continentPopulation = AdditionalReports.getPopulationOfContinent(con, continent);
+        // Long continentPopulation = Sql.getPopulationOfContinent(con, continent);
         // System.out.println("Population of " + continent + " is " + continentPopulation);
 
         // Report 3:
         // Get population of a region
         // Set region to get population of
         //String region = "Southern Europe";
-        //Long regionPopulation = AdditionalReports.getPopulationOfRegion(con, region);
+        //Long regionPopulation = Sql.getPopulationOfRegion(con, region);
         //System.out.println("Population of " + region + " is " + regionPopulation);
 
         // Report 4:
         // Get the population of a country
         // Set the country code
         //String code = "GBR";
-        //Long countryPopulation = AdditionalReports.getCountryPopulation(con, code);
+        //Long countryPopulation = Sql.getCountryPopulation(con, code);
         //System.out.println("The population of " + code + " is " + countryPopulation);
 
         // Report 5:
         // Get the population of a district
         // Set the district
         //String district = "England";
-        //Long districtPopulation = AdditionalReports.getDistrictPopulation(con, district);
+        //Long districtPopulation = Sql.getDistrictPopulation(con, district);
         //System.out.println("The population of " + district + " is " + districtPopulation);
 
         // Report 6:
         // Get the population of a city
         // Set the city name
         //String cityName = "London";
-        //Long cityPopulation = AdditionalReports.getCityPopulation(con, cityName);
+        //Long cityPopulation = Sql.getCityPopulation(con, cityName);
         //System.out.println("The population of " + cityName + " is " + cityPopulation);
 
-        // Population of people... reports //
+        /************** POPULATION FUNCTIONS ***************** Author Cameron */
         // Report 1:
         // The population of people, people living in cities, and people not living in cities in each continent.
         // Antarctica is not shown as the population is set to 0 in the database //
-        //ArrayList<PopulationData> continentPopulationDataList = AdditionalReports.getPopulationDataContinent(con);
+        //ArrayList<PopulationData> continentPopulationDataList = Sql.getPopulationDataContinent(con);
         //Reports.printPopulationDataContinent(continentPopulationDataList);
 
         // Report 2:
         // The population of people, people living in cities, and people not living in cities in each region.
-        //ArrayList<PopulationData> regionPopulationDataList = AdditionalReports.getPopulationDataRegion(con);
+        //ArrayList<PopulationData> regionPopulationDataList = Sql.getPopulationDataRegion(con);
         //Reports.printPopulationDataRegion(regionPopulationDataList);
 
         // Report 3:
         // The population of people, people living in cities, and people not living in cities in each country.
-        //ArrayList<PopulationData> countryPopulationDataList = AdditionalReports.getPopulationDataCountry(con);
+        //ArrayList<PopulationData> countryPopulationDataList = Sql.getPopulationDataCountry(con);
         //Reports.printPopulationDataCountry(countryPopulationDataList);
 
         // Disconnect from database
