+++ src/main/java/com/napier/sem/sql/Sql.java
@@ -2,6 +2,7 @@ package com.napier.sem.sql;
 
 import com.napier.sem.world.City;
 import com.napier.sem.world.Country;
+import com.napier.sem.world.PopulationData;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -11,11 +12,14 @@ import java.util.ArrayList;
 
 /**
  * Project Authors: Colin, Cameron, Luke, Del
- * Date last modified: 21/02/2021
+ * Date last modified: 22/02/2021
  * Purpose of class: This class handles the SQL queries to the database.
+ * Last modified by: Colin
  */
 
 public class Sql {
+
+    /************** COUNTRIES QUERIES ***************** Author Luke */
 
     /**
      * Gets all the countries in the world from
@@ -23,7 +27,7 @@ public class Sql {
      * @param con Holds the connection to the SQL database
      * @return All countries and their populations in the world in order of largest to smallest in an ArrayList
      * Report 1 */
-    public ArrayList<Country> getAllCountriesInWorld(Connection con)
+    public static ArrayList<Country> getAllCountriesInWorld(Connection con)
     {
         try
         {
@@ -58,13 +62,13 @@ public class Sql {
     }
 
     /**
-     * Gets all the countries in a continet from
+     * Gets all the countries in a continent from
      * the database and returns the countries object filled with data
      * @param con Holds the connection to the SQL database
      * @param continent Holds the string of a specified continent
      * @return All countries and their populations in a continent in order of largest to smallest in an ArrayList
      * Report 2 */
-    public ArrayList<Country> getAllCountriesInContinent(Connection con, String continent)
+    public static ArrayList<Country> getAllCountriesInContinent(Connection con, String continent)
     {
         try
         {
@@ -110,10 +114,10 @@ public class Sql {
      * Gets all the countries in a continent from
      * the databse and returns the countries object filled with data
      * @param con Holds the connection to the SQL database
-     * @param region Holds the of a specified region
+     * @param region Holds the string of a specified region
      * @return All countries and their populations in a region in order of largest to smallest in an ArrayList
      * Report 3 */
-    public ArrayList<Country> getAllCountriesByRegion(Connection con, String region)
+    public static ArrayList<Country> getAllCountriesByRegion(Connection con, String region)
     {
         try
         {
@@ -162,7 +166,7 @@ public class Sql {
      * @param limit Holds the a number to limit the amount of results
      * @return The top N countries and their populations in the world in order of largest to smallest in an ArrayList
      * Report 4 */
-    public ArrayList<Country> getTopPopulatedCountriesInWorld(Connection con, int limit)
+    public static ArrayList<Country> getTopPopulatedCountriesInWorld(Connection con, int limit)
     {
         try
         {
@@ -207,11 +211,11 @@ public class Sql {
      * Gets top populated countries in a continent from
      * the database and returns the countries object filled with data
      * @param con Holds the connection to the SQL database
-     * @param continent
+     * @param continent Holds the string of a specified continent
      * @param limit Holds the a number to limit the amount of results
      * @return The top N countries and their populations in the continent in order of largest to smallest in an ArrayList
      * Report 5 */
-    public ArrayList<Country> getTopPopulatedCountriesInContinent(Connection con, String continent, int limit)
+    public static ArrayList<Country> getTopPopulatedCountriesInContinent(Connection con, String continent, int limit)
     {
         try
         {
@@ -259,11 +263,11 @@ public class Sql {
      * Gets top populated countries in a region from
      * the database and returns the countries object fulled with data
      * @param con Holds the connection to the SQL database
-     * @param region
+     * @param region Holds the string of a specified region
      * @param limit Holds the a number to limit the amount of results
      * @return The top N countries and their populations in the region in order of largest to smallest in an ArrayList
      * report 6 */
-    public ArrayList<Country> getTopPopulatedCountriesInRegion(Connection con, String region, int limit)
+    public static ArrayList<Country> getTopPopulatedCountriesInRegion(Connection con, String region, int limit)
     {
         try
         {
@@ -307,8 +311,13 @@ public class Sql {
 
     }
 
-    /** Gets all the cities in the world from
+    /************** CITIES QUERIES ***************** Author Colin */
+
+    /**
+     * Gets all the cities in the world from
      * the database and returns the cities object filled with data.
+     * @param con Holds the connection to the SQL database
+     * @return All cities in the world object data
      * Report 7 */
     public static ArrayList<City> getAllCitiesInWorld(Connection con)
     {
@@ -344,10 +353,14 @@ public class Sql {
         }
     }
 
-    /** Gets all the cities in a chosen continent from
+    /**
+     * Gets all the cities in a chosen continent from
      * the database and returns the cities object filled with data.
+     * @param con Holds the connection to the SQL database
+     * @param continent Holds the string of a specified continent
+     * @return All the cities in a chosen continent object
      * Report 8 */
-    public ArrayList<City> getAllCitiesInContinent(Connection con, String continent)
+    public static ArrayList<City> getAllCitiesInContinent(Connection con, String continent)
     {
 
         try
@@ -385,10 +398,14 @@ public class Sql {
         }
     }
 
-    /** Gets all the cities in a chosen region from
+    /**
+     * Gets all the cities in a chosen region from
      * the database and returns the cities object filled with data.
+     * @param con Holds the connection to the SQL database
+     * @param region Holds the string of a specified region
+     * @return all cities in chosen region object
      * Report 9 */
-    public ArrayList<City> getAllCitiesInRegion(Connection con, String region)
+    public static ArrayList<City> getAllCitiesInRegion(Connection con, String region)
     {
 
         try
@@ -425,10 +442,14 @@ public class Sql {
         }
     }
 
-    /** Gets all the cities in a chosen country from
+    /**
+     * Gets all the cities in a chosen country from
      * the database and returns the cities object filled with data.
+     * @param con Holds the connection to the SQL database
+     * @param country Holds the string of a specified country
+     * @return object with all cities in chosen country
      * Report 10 */
-    public ArrayList<City> getAllCitiesInCountry(Connection con, String country)
+    public static ArrayList<City> getAllCitiesInCountry(Connection con, String country)
     {
 
         try
@@ -466,10 +487,14 @@ public class Sql {
         }
     }
 
-    /** Gets all the cities in a chosen district from
+    /**
+     * Gets all the cities in a chosen district from
      * the database and returns the cities object filled with data.
+     * @param con Holds the connection to the SQL database
+     * @param district Holds the string of a specified district
+     * @return all cities in a chosen district object
      * Report 11 */
-    public ArrayList<City> getAllCitiesInDistrict(Connection con, String district)
+    public static ArrayList<City> getAllCitiesInDistrict(Connection con, String district)
     {
 
         try
@@ -507,10 +532,14 @@ public class Sql {
         }
     }
 
-    /** Gets all the N populated cities in the world from
+    /**
+     * Gets all the N populated cities in the world from
      * the database and returns the cities object filled with data.
+     * @param con Holds the connection to the SQL database
+     * @param limit The number of results from the query
+     * @return Top populated cities in the world object
      * Report 12 */
-    public ArrayList<City> getTopPopulatedCitiesInWorld(Connection con, int limit)
+    public static ArrayList<City> getTopPopulatedCitiesInWorld(Connection con, int limit)
     {
 
         try
@@ -549,10 +578,15 @@ public class Sql {
         }
     }
 
-    /** Gets all the N populated cities in a chosen continent from
+    /**
+     * Gets all the N populated cities in a chosen continent from
      * the database and returns the cities object filled with data.
+     * @param con Holds the connection to the SQL database
+     * @param limit The number of results from the query
+     * @param continent Holds the string of a specified continent
+     * @return Top populated cities in a continent object
      * Report 13 */
-    public ArrayList<City> getTopPopulatedCitiesInContinent(Connection con, int limit, String continent)
+    public static ArrayList<City> getTopPopulatedCitiesInContinent(Connection con, int limit, String continent)
     {
 
         try
@@ -592,10 +626,15 @@ public class Sql {
         }
     }
 
-    /** Gets all the N populated cities in a chosen region from
+    /**
+     * Gets all the N populated cities in a chosen region from
      * the database and returns the cities object filled with data.
+     * @param con Holds the connection to the SQL database
+     * @param limit The number of results from the query
+     * @param region Holds the string of a specified region
+     * @return Top populated cities in a region object
      * Report 14 */
-    public ArrayList<City> getTopPopulatedCitiesInRegion(Connection con, int limit, String region)
+    public static ArrayList<City> getTopPopulatedCitiesInRegion(Connection con, int limit, String region)
     {
 
         try
@@ -634,10 +673,15 @@ public class Sql {
         }
     }
 
-    /** Gets all the N populated cities in a chosen country from
+    /**
+     * Gets all the N populated cities in a chosen country from
      * the database and returns the cities object filled with data.
+     * @param con Holds the connection to the SQL database
+     * @param limit The number of results from the query
+     * @param country Holds the string of a specified country
+     * @return Top populated cities in a country object
      * Report 15 */
-    public ArrayList<City> getTopPopulatedCitiesInCountry(Connection con, int limit, String country)
+    public static ArrayList<City> getTopPopulatedCitiesInCountry(Connection con, int limit, String country)
     {
 
         try
@@ -678,10 +722,15 @@ public class Sql {
         }
     }
 
-    /** Gets all the N populated cities in a chosen district from
+    /**
+     * Gets all the N populated cities in a chosen district from
      * the database and returns the cities object filled with data.
+     * @param con Holds the connection to the SQL database
+     * @param limit The number of results from the query
+     * @param district Holds the string of a specified district
+     * @return Top populated cities in a district object
      * Report 16 */
-    public ArrayList<City> getTopPopulatedCitiesInDistrict(Connection con, int limit, String district)
+    public static ArrayList<City> getTopPopulatedCitiesInDistrict(Connection con, int limit, String district)
     {
 
         try
@@ -716,6 +765,312 @@ public class Sql {
         {
             System.out.println(e.getMessage());
             System.out.println("Failed to get city district details");
+            return null;
+        }
+    }
+
+    /************** CAPITAL CITIES QUERIES ***************** Author Del */
+
+
+
+
+    /************** POPULATION QUERIES ***************** Author Cameron */
+
+    /**
+     * Method to return a list of population data for each continent
+     * @param con Holds the connection to the SQL database
+     * @return The total population, population living in cities & population not living in cities of each continent in an ArrayList,
+     * or null if there is an error.
+     * Report 23 */
+    public static ArrayList<PopulationData> getPopulationDataContinent(Connection con) {
+        try {
+            // Create SQL statement
+            Statement peopleInCitiesStatement = con.createStatement();
+            // Create string for SQL statement
+            String citiesPopulation = "SELECT country.continent, SUM(city.population) as citiesPopulation " +
+                    "FROM city " +
+                    "JOIN country ON CountryCode=country.code " +
+                    "GROUP BY country.continent; ";
+            // Execute SQL statement
+            ResultSet rset = peopleInCitiesStatement.executeQuery(citiesPopulation);
+            // ArrayList to store all population data required
+            ArrayList<PopulationData> allPopulationData = new ArrayList<>();
+            // Return population data of each continent
+            // Check something is returned
+            while(rset.next()) {
+                PopulationData popData = new PopulationData();
+                popData.identifier = rset.getString("country.continent");
+                popData.population = getPopulationOfContinent(con, popData.identifier);
+                popData.populationInCities = rset.getLong("citiesPopulation");
+                popData.populationNotInCities = popData.population - popData.populationInCities;
+                allPopulationData.add(popData);
+            }
+            return allPopulationData;
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            System.out.println("Failed to get the requested population data.");
+            return null;
+        }
+    }
+
+    /**
+     * Method to return a list of population data for each region
+     * @param con Holds the connection to the SQL database
+     * @return The total population, population living in cities & population not living in cities of each region in an ArrayList,
+     * or null if there is an error.
+     * Report 24 */
+    public static ArrayList<PopulationData> getPopulationDataRegion(Connection con) {
+        try {
+            // Create SQL statement
+            Statement peopleInCitiesStatement = con.createStatement();
+            // Create string for SQL statement
+            String citiesPopulation = "SELECT country.region, SUM(city.population) as citiesPopulation " +
+                    "FROM city " +
+                    "JOIN country ON CountryCode=country.code " +
+                    "GROUP BY country.region; ";
+            // Execute SQL statement
+            ResultSet rset = peopleInCitiesStatement.executeQuery(citiesPopulation);
+            // ArrayList to store all population data required
+            ArrayList<PopulationData> allPopulationData = new ArrayList<>();
+            // Return population data of each region
+            // Check something is returned
+            while(rset.next()) {
+                PopulationData popData = new PopulationData();
+                popData.identifier = rset.getString("country.region");
+                popData.population = getPopulationOfRegion(con, popData.identifier);
+                popData.populationInCities = rset.getLong("citiesPopulation");
+                popData.populationNotInCities = popData.population - popData.populationInCities;
+                allPopulationData.add(popData);
+            }
+            return allPopulationData;
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            System.out.println("Failed to get the requested population data.");
+            return null;
+        }
+    }
+
+    /**
+     * Method to return a list of population data for each country
+     * @param con Holds the connection to the SQL database
+     * @return The total population, population living in cities & population not living in cities of each country in an ArrayList,
+     * or null if there is an error.
+     * Report 25 */
+    public static ArrayList<PopulationData> getPopulationDataCountry(Connection con) {
+        try {
+            // Create SQL statement
+            Statement peopleInCitiesStatement = con.createStatement();
+            // Create string for SQL statement
+            String citiesPopulation = "SELECT country.code, country.name, SUM(city.population) as citiesPopulation " +
+                    "FROM city " +
+                    "JOIN country ON CountryCode=country.code " +
+                    "GROUP BY country.name, country.code; ";
+            // Execute SQL statement
+            ResultSet rset = peopleInCitiesStatement.executeQuery(citiesPopulation);
+            // ArrayList to store all population data required
+            ArrayList<PopulationData> allPopulationData = new ArrayList<>();
+            // Return population data of each country
+            // Check something is returned
+            while(rset.next()) {
+                PopulationData popData = new PopulationData();
+                popData.code = rset.getString("country.code");
+                popData.identifier = rset.getString("country.name");
+                popData.population = getCountryPopulation(con, popData.code);
+                popData.populationInCities = rset.getLong("citiesPopulation");
+                popData.populationNotInCities = popData.population - popData.populationInCities;
+                allPopulationData.add(popData);
+            }
+            return allPopulationData;
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            System.out.println("Failed to get the requested population data.");
+            return null;
+        }
+    }
+
+
+    /************** ADDITIONAL QUERIES ***************** Author Cameron */
+
+    /**
+     * Method to return the population of the world
+     * @param con Holds the connection to the SQL server
+     * @return The population of the world or null if there is an error
+     * Report 26 */
+    public static Long getWorldPopulation(Connection con) {
+        try {
+            // Create SQL statement
+            Statement populationStatement = con.createStatement();
+            // Create string for SQL statement
+            String worldPopulation = "SELECT SUM(population) as worldPopulation " +
+                    "FROM country;";
+            // Execute SQL statement
+            ResultSet rset = populationStatement.executeQuery(worldPopulation);
+            // Return world population if calculation is valid
+            // Check something is returned
+            if(rset.next()) {
+                return rset.getLong("worldPopulation");
+            } else
+                return null;
+
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            System.out.println("Failed to calculate the population of the world.");
+            return null;
+        }
+    }
+
+    /**
+     * Method to return the population of a given continent
+     * @param con Holds the connection to the SQL server
+     * @param continent Contains the continent to get the population for
+     * @return The population of the continent or null if there is an error
+     * Report 27 */
+    public static Long getPopulationOfContinent(Connection con, String continent) {
+        try {
+            // Create string for SQL statement
+            String continentPopulation = "SELECT SUM(population) as continentPopulation " +
+                    "FROM country " +
+                    "WHERE continent = ?;";
+
+            // Create prepared statement with SQL statement
+            PreparedStatement preparedStatement = con.prepareStatement(continentPopulation);
+            // Set SQL statement ? to continent parameter
+            preparedStatement.setString(1, continent);
+            // Execute SQL statement
+            ResultSet rset = preparedStatement.executeQuery();
+            // Return population of continent if calculation is valid
+            // Check something is returned
+            if(rset.next()) {
+                return rset.getLong("continentPopulation");
+            } else
+                return null;
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            System.out.println("Failed to calculate the population of the continent " + continent);
+            return null;
+        }
+    }
+
+    /**
+     * Method to return the population of a specified region
+     * @param con Holds the connection to the SQL server
+     * @param region Contains the region to get the population for
+     * @return The population of the region or null if there is an error
+     * Report 28 */
+    public static Long getPopulationOfRegion(Connection con, String region) {
+        try {
+            // Create string for SQL statement
+            String regionPopulation = "SELECT SUM(population) as regionPopulation " +
+                    "FROM country " +
+                    "WHERE region = ?;";
+            // Create prepared statement with SQL statement
+            PreparedStatement preparedStatement = con.prepareStatement(regionPopulation);
+            // Set SQL statement ? to continent parameter
+            preparedStatement.setString(1, region);
+            // Execute SQL statement
+            ResultSet rset = preparedStatement.executeQuery();
+            // Return population of continent if calculation is valid
+            // Check something is returned
+            if(rset.next()) {
+                return rset.getLong("regionPopulation");
+            } else
+                return null;
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            System.out.println("Failed to calculate the population of the region " + region);
+            return null;
+        }
+    }
+
+    /**
+     * Method to return the population for a specified country
+     * @param con Holds the connection to the SQL server
+     * @param code Contains the unique country code used to retrieve the correct population
+     * @return The population of a country or null if there is an error
+     * Report 29 */
+    public static Long getCountryPopulation(Connection con, String code) {
+        try {
+            // Create string for SQL statement
+            String countryPopulation = "SELECT name, population " +
+                    "FROM country " +
+                    "WHERE code = ?;";
+            // Create prepared statement with SQL statement
+            PreparedStatement preparedStatement = con.prepareStatement(countryPopulation);
+            // Set SQL statement ? to continent parameter
+            preparedStatement.setString(1, code);
+            // Execute SQL statement
+            ResultSet rset = preparedStatement.executeQuery();
+            // Return population of specified country + country name
+            // Check something is returned
+            if(rset.next()) {
+                return rset.getLong("population");
+            } else
+                return null;
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            System.out.println("Failed to get the population of the country with the code " + code);
+            return null;
+        }
+    }
+
+    /**
+     * Method to return the population of a specified district
+     * @param con Holds the connection to the SQL server
+     * @param district Contains the district to get the population for
+     * @return The population of a given district or null if there is an error
+     * Report 30 */
+    public static Long getDistrictPopulation(Connection con, String district) {
+        try {
+            // Create string for SQL statement
+            String districtPopulation = "SELECT SUM(population) as districtPopulation " +
+                    "FROM city " +
+                    "WHERE district = ?;";
+            // Create prepared statement with SQL statement
+            PreparedStatement preparedStatement = con.prepareStatement(districtPopulation);
+            // Set SQL statement ? to continent parameter
+            preparedStatement.setString(1, district);
+            // Execute SQL statement
+            ResultSet rset = preparedStatement.executeQuery();
+            // Return population of specified district
+            // Check something is returned
+            if(rset.next()) {
+                return rset.getLong("districtPopulation");
+            } else
+                return null;
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            System.out.println("Failed to get the population of the district.");
+            return null;
+        }
+    }
+
+    /**
+     * Method to return the population of a specified city
+     * @param con Holds the connection to the SQL server
+     * @param cityName Contains the city to get the population for
+     * @return The population of a given city or null if there is an error
+     * Report 31 */
+    public static Long getCityPopulation(Connection con, String cityName) {
+        try {
+            // Create string for SQL statement
+            String cityPopulation = "SELECT population " +
+                    "FROM city " +
+                    "WHERE name = ?;";
+            // Create prepared statement with SQL statement
+            PreparedStatement preparedStatement = con.prepareStatement(cityPopulation);
+            // Set SQL statement ? to cityName parameter
+            preparedStatement.setString(1, cityName);
+            // Execute SQL statement
+            ResultSet rset = preparedStatement.executeQuery();
+            // Return population of specified city
+            // Check something is returned
+            if(rset.next()) {
+                return rset.getLong("population");
+            } else
+                return null;
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            System.out.println("Failed to get the population of " + cityName);
             return null;
         }
     }
